reorganize components folders
