+++ components/Layout/PreviewBar/index.js
