+++ components/Pieces/Button/index.js
@@ -1,6 +1,6 @@
 import classNames from 'classnames'
 import PropTypes from 'prop-types'
-import PrismicLink from 'components/PrismicLink'
+import PrismicLink from 'components/Pieces/PrismicLink'
 import styles from './index.module.scss'
 import prismicLinkShape from 'shapes/prismic/link'
 const Button = ({ className, link, children }) => {
