+++ components/Templates/Page/index.js
@@ -1,8 +1,8 @@
 import PropTypes from 'prop-types'
 import ErrorPage from 'next/error'
-import Layout from 'components/Layout'
-import Slices from 'components/Slices'
-import headerShape from 'components/Header/shape'
+import Layout from 'components/Layout/_Layout'
+import Slices from 'components/Slices/_Slices'
+import headerShape from 'components/Layout/Header/shape'
 import { useRouter } from 'next/router'
 import pageShape from './shape'
